feat: export from_url to acs_mixins
